refactor(code)!: update stylelint to version 16
BREAKING CHANGE:
Require stylelint v16
