+++ .github/workflows/test.yml
@@ -13,9 +13,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest]
-        node: [16, 18, 20]
-        # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        node: [18, 20, current]
 
     runs-on: ${{ matrix.os }}
