+++ .gitignore
@@ -1,7 +1,7 @@
 # https://git-scm.com/docs/gitignore
 # Folders to ignore
 node_modules
-.nvmrc
+archive
 
 # To do Tasks
 .todo
