+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "gulp-stylelint-esm",
-  "version": "1.1.0",
-  "description": "Gulp plugin for running Stylelint results through various reporters.",
+  "version": "2.0.0-dev",
+  "description": "Gulp plugin for stylelint with ES module support.",
   "author": {
     "name": "Adorade",
     "email": "office.adorade@gmail.com",
@@ -13,13 +13,13 @@
     "url": "https://github.com/adorade/gulp-stylelint-esm.git"
   },
   "type": "module",
-  "exports": "./src/index.js",
+  "exports": "./src/index.mjs",
   "files": [
-    "/src/*.js"
+    "/src/*.mjs"
   ],
   "keywords": [
-    "gulp",
-    "plugin",
+    "gulpplugin",
+    "gulpfriendly",
     "stylelint",
     "postcss",
     "css",
@@ -27,13 +27,13 @@
     "esm"
   ],
   "scripts": {
-    "lint": "eslint \"{src,test}/**/*.js\"",
-    "tape": "tape \"test/*.spec.js\"",
+    "lint": "eslint {src,test}/**/*.{mjs,js}",
+    "tape": "tape test/*.spec.js",
     "test": "yarn run lint && yarn run tape",
     "prepublishOnly": "yarn test"
   },
   "peerDependencies": {
-    "stylelint": "^14.0.0 || ^15.0.0"
+    "stylelint": "^16.0.2"
   },
   "dependencies": {
     "ansi-colors": "4.1.3",
@@ -42,20 +42,20 @@
     "source-map": "0.7.4"
   },
   "devDependencies": {
-    "@babel/core": "7.23.3",
+    "@babel/core": "7.23.6",
     "@babel/eslint-parser": "7.23.3",
-    "eslint": "8.54.0",
+    "eslint": "8.55.0",
     "eslint-config-stylelint": "20.0.0",
     "gulp": "4.0.2",
     "gulp-clean-css": "4.3.0",
     "gulp-concat": "2.6.1",
     "gulp-rename": "2.0.0",
     "sinon": "17.0.1",
-    "stylelint": "15.11.0",
+    "stylelint": "16.0.2",
     "tape": "5.7.2"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=18.12.0"
   },
   "license": "MIT"
 }
