+++ test/fixtures/invalid.css
@@ -1,3 +1,3 @@
 .foo {
-  color: #FFF;
+  color: #ffffff;
 }
