+++ test/reporter-factory.spec.js
@@ -1,5 +1,5 @@
 /*!
- * Gulp Stylelint (v1.1.0): test/reporter-factory.spec.js
+ * Gulp Stylelint (v2.0.0-dev): test/reporter-factory.spec.js
  * Copyright (c) 2023 Adorade (https://github.com/adorade/gulp-stylelint-esm)
  * License under MIT
  * ========================================================================== */
@@ -8,7 +8,7 @@ import fancyLog from 'fancy-log';
 import { stub } from 'sinon';
 import test from 'tape';
 
-import reporterFactory from '../src/reporter-factory.js';
+import { reporterFactory } from '../src/reporter-factory.mjs';
 
 test('reporter factory should return a function', t => {
   t.plan(1);
